static: Add common CSS for images, videos, embeds
Compared to the previous approach of using inline styles everywhere,
this provides less style duplication, better mobile support, and
encourages the use of semantic tags instead of just <div>s and <p>s.

The layout is simpler to understand without all the inline styles
and <br>s and also easier to maintain.

This also enables styling to be more consistent for different kinds
of embeds (images, videos, iframe embeds such as gifs).