Fix media-block having 0 height with only floating children
The media-block element has 0 height when it contains only floating
children, has floating elements do not induce height on their parent.
The result is text sitting to the right of the media when it shouldn't.
Using overflow: hidden on it fixes this by establishing a new block
formatting context, as explained here:
https://www.w3.org/TR/CSS21/visuren.html#block-formatting