Fix title and image in Discord previews (and possibly other sites)
og:title was hardcoded to "Dolphin Emulator" instead of being set to
the article title or the page name. Let's just remove it and have
FB/Twitter/Discord/etc. fall back to using <title>.
(We can't use the "title" block to generate the og:title value because
template blocks can only be used once.)

og:type was set to an invalid value.

og:image is a dead link, and it was pointing to an image hosted on
Facebook's CDN for some reason (???). Just remove it and let crawlers
figure out what image to show (usually the article banner).

This should fix article previews showing
"Dolphin Emulator - Dolphin Emulator" with no image in Discord.