update: migrate macos users to macos-universal
This makes sure existing users will not be left behind on the Intel
macOS builds after the old macOS VM is shut down.

An alternative approach would have been to add some special logic
in Dolphin's auto-update code to request a macos-universal target
manifest on macos builds (instead of defaulting to the current
platform). The downside to this approach is that users who are
on older builds would have to update twice: the first time to a
version of Dolphin that has the aforementioned special logic, and
a second time to actually switch to a macos-universal build.