downloads: Handle duplicate version names
If there is more than one version on a branch with a given shortrev,
get_object_or_404 would raise a MultipleObjectsReturned exception. It is
better just to pick one. Since these duplicate versions come from force
pushes, the oldest should be the original and thus the most useful one
to pick.